link-parser.c: Move the screen width handling to parser-utilities.c
